+++ src/contracts/interfaces/IBridgeDistribution.sol
@@ -2,11 +2,11 @@
 pragma solidity 0.8.29;
 
 interface IBribeDistribution {
-    function _deposit(uint amount, uint tokenId) external;
-    function _withdraw(uint amount, uint tokenId) external;
-    function getRewardForOwner(uint tokenId, address[] memory tokens) external;
-    function notifyRewardAmount(address token, uint amount) external;
-    function left(address token) external view returns (uint);
-    function recoverERC20(address tokenAddress, uint256 tokenAmount) external ;
+    function _deposit(uint256 amount, uint256 tokenId) external;
+    function _withdraw(uint256 amount, uint256 tokenId) external;
+    function getRewardForOwner(uint256 tokenId, address[] memory tokens) external;
+    function notifyRewardAmount(address token, uint256 amount) external;
+    function left(address token) external view returns (uint256);
+    function recoverERC20(address tokenAddress, uint256 tokenAmount) external;
     function setOwner(address _owner) external;
 }
